try/catch blocks added
These were added to make sure there is a big bang, to alert the user that something is not right here.

Properties cannot work with NULL (neither key nor value).
Instead of adding simple `!= null` checks, exceptions are thrown, so the user (hopefully) sees that the error is on him or one of the plugins.